Add: system for terminal ClI commands with click
