+++ tests/test2.py
@@ -1,0 +1,4 @@
+import pyfiglet
+
+ascii_banner = pyfiglet.figlet_format("SYSTOOLS")
+print(ascii_banner)
